refactor formatQueryInfo into fauna module
